backend/pyproject.toml
[tool.poetry]
name = "current-paychecks"
version = "0.1.0"
description = "REST-сервис просмотра текущей зарплаты и даты следующего повышения"
authors = ["exp-ext <ext77@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.6"
fastapi = "^0.95.2"

asyncpg = "^0.28.0"
SQLAlchemy = "^2.0.0"
alembic = "^1.11.1"
databases = {version = "^0.4.2", extras = ["postgresql"]}
uvicorn = "^0.22.0"

python-dotenv = "^1.0.0"

passlib = "^1.7.4"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"

pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
